Expanded queues object to include more stats.
Added a skipped list
Added a played list
Added a 'repeats' variable.  This will be used to determing if a queue
should be allowed to contain repeats.  More granularity can be added later
like if a song that has already been played can be added back so that only the
current queue is all unique etc.

Signed-off-by: Tom Schutte <tjschutte@wisc.edu>